Matt's changes to N8 presentation
Changes the slides for the example used to explain why binary comparisons are bad, as well as the example that introduces ARIMA modelling. Also splits the interview quotes so each quote is on a separate slide, re-arranges the conclusion slides and adds a QR code and bitly link in place of the long URL used previously.